-renamed resolveStylesAndClasses to resolveStyles and merged two functions into one -updated tests
