+++ packages/react-bindings/test/styles/resolveStyles-test.ts
@@ -1,22 +1,11 @@
 import {
   ComponentSlotStylesPrepared,
-  ComponentStyleFunctionParam,
   emptyTheme,
   ICSSInJSStyle,
-  ThemePrepared,
 } from '@fluentui/styles'
-import resolveStylesAndClasses from '../../src/styles/resolveStylesAndClasses'
+import { PrimitiveProps } from '@fluentui/react-bindings'
+import resolveStyles, { ResolveStylesInput } from '../../src/styles/resolveStyles'
 
-const styleParam: ComponentStyleFunctionParam = {
-  disableAnimations: false,
-  displayName: 'Test',
-  props: {},
-  rtl: false,
-  theme: emptyTheme,
-  variables: {
-    color: 'red',
-  },
-}
 
 const componentStyles: ComponentSlotStylesPrepared<{}, { color: string }> = {
   root: ({ variables: v }): ICSSInJSStyle => ({
@@ -24,18 +13,52 @@ const componentStyles: ComponentSlotStylesPrepared<{}, { color: string }> = {
   }),
 }
 
-const theme: ThemePrepared = emptyTheme
+const resolveStylesParam = (options?: {
+  displayName?: string
+  renderStyles?: (styles: ICSSInJSStyle) => string
+  cacheEnabled?: boolean
+  stylesProps?: PrimitiveProps
+}): ResolveStylesInput => {
+  const {
+    displayName = 'Test',
+    cacheEnabled = false,
+    stylesProps = {},
+    renderStyles = undefined,
+  } = options || {}
+
+  return {
+    theme: {
+      ...emptyTheme,
+      componentStyles: {
+        [displayName]: componentStyles,
+      },
+    },
+    displayName,
+    props: {},
+    resolvedVariables: {
+      color: 'red',
+    },
+    rtl: false,
+    disableAnimations: false,
+    renderer: {
+      renderRule: () => '',
+    },
+    cacheEnabled,
+    stylesProps,
+    renderStyles,
+  }
+}
 
 describe('resolveStylesAndClasses', () => {
   test('resolves styles', () => {
-    const { resolvedStyles } = resolveStylesAndClasses(componentStyles, styleParam, () => '')
+    const { resolvedStyles } = resolveStyles(resolveStylesParam())
 
     expect(resolvedStyles.root).toMatchObject({ color: 'red' })
   })
 
   test('caches resolved styles', () => {
     spyOn(componentStyles, 'root').and.callThrough()
-    const { resolvedStyles } = resolveStylesAndClasses(componentStyles, styleParam, () => '')
+    const { resolvedStyles } = resolveStyles(resolveStylesParam())
 
     expect(resolvedStyles.root).toMatchObject({ color: 'red' })
     expect(componentStyles.root).toHaveBeenCalledTimes(1)
@@ -45,7 +68,7 @@ describe('resolveStylesAndClasses', () => {
 
   test('does not render classes if not fetched', () => {
     const renderStyles = jest.fn()
-    const { resolvedStyles } = resolveStylesAndClasses(componentStyles, styleParam, renderStyles)
+    const { resolvedStyles } = resolveStyles(resolveStylesParam({ renderStyles }))
 
     expect(resolvedStyles.root).toMatchObject({ color: 'red' })
     expect(renderStyles).not.toBeCalled()
@@ -53,7 +76,7 @@ describe('resolveStylesAndClasses', () => {
 
   test('renders classes when slot classes getter is accessed', () => {
     const renderStyles = jest.fn().mockReturnValue('a')
-    const { classes } = resolveStylesAndClasses(componentStyles, styleParam, renderStyles)
+    const { classes } = resolveStyles(resolveStylesParam({ renderStyles }))
 
     expect(classes['__root']).toBeDefined()
     expect(renderStyles).toHaveBeenCalledWith({ color: 'red' })
@@ -61,7 +84,7 @@ describe('resolveStylesAndClasses', () => {
 
   test('caches rendered classes', () => {
     const renderStyles = jest.fn().mockReturnValue('a')
-    const { classes } = resolveStylesAndClasses(componentStyles, styleParam, renderStyles)
+    const { classes } = resolveStyles(resolveStylesParam({ renderStyles }))
 
     expect(classes['__root']).toBeDefined()
     expect(renderStyles).toHaveBeenCalledWith({ color: 'red' })
@@ -71,24 +94,9 @@ describe('resolveStylesAndClasses', () => {
 
   test('caches resolved styles for no props', () => {
     spyOn(componentStyles, 'root').and.callThrough()
-    const { resolvedStyles } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      () => '',
-      true,
-      'Test',
-      theme,
-      {},
-    )
-    const { resolvedStyles: secondResolvedStyles } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      () => '',
-      true,
-      'Test',
-      theme,
-      {},
-    )
+    const options = resolveStylesParam({ cacheEnabled: true})
+    const { resolvedStyles } = resolveStyles(options)
+    const { resolvedStyles: secondResolvedStyles } = resolveStyles(options)
 
     expect(resolvedStyles.root).toMatchObject({ color: 'red' })
     expect(componentStyles.root).toHaveBeenCalledTimes(1)
@@ -98,24 +106,13 @@ describe('resolveStylesAndClasses', () => {
 
   test('caches classes for no props', () => {
     const renderStyles = jest.fn().mockReturnValue('a')
-    const { classes } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
+    const options = resolveStylesParam({
       renderStyles,
-      true,
-      'Test1',
-      theme,
-      {},
-    )
-    const { classes: secondClasses } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      renderStyles,
-      true,
-      'Test1',
-      theme,
-      {},
-    )
+      cacheEnabled: true,
+      displayName: 'Test1',
+    })
+    const { classes } = resolveStyles(options)
+    const { classes: secondClasses } = resolveStyles(options)
 
     expect(classes['__root']).toBeDefined()
     expect(renderStyles).toHaveBeenCalledWith({ color: 'red' })
@@ -125,24 +122,13 @@ describe('resolveStylesAndClasses', () => {
 
   test('caches resolved styles for the same props', () => {
     spyOn(componentStyles, 'root').and.callThrough()
-    const { resolvedStyles } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      () => '',
-      true,
-      'Test2',
-      theme,
-      { primary: true },
-    )
-    const { resolvedStyles: secondResolvedStyles } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      () => '',
-      true,
-      'Test2',
-      theme,
-      { primary: true },
-    )
+    const options = resolveStylesParam({
+      cacheEnabled: true,
+      displayName: 'Test2',
+      stylesProps: { primary: true },
+    })
+    const { resolvedStyles } = resolveStyles(options)
+    const { resolvedStyles: secondResolvedStyles } = resolveStyles(options)
 
     expect(resolvedStyles.root).toMatchObject({ color: 'red' })
     expect(componentStyles.root).toHaveBeenCalledTimes(1)
@@ -152,24 +138,14 @@ describe('resolveStylesAndClasses', () => {
 
   test('caches classes for the same props', () => {
     const renderStyles = jest.fn().mockReturnValue('a')
-    const { classes } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
+    const options = resolveStylesParam({
       renderStyles,
-      true,
-      'Test3',
-      theme,
-      { primary: true },
-    )
-    const { classes: secondClasses } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      renderStyles,
-      true,
-      'Test3',
-      theme,
-      { primary: true },
-    )
+      cacheEnabled: true,
+      displayName: 'Test3',
+      stylesProps: { primary: true },
+    })
+    const { classes } = resolveStyles(options)
+    const { classes: secondClasses } = resolveStyles(options)
 
     expect(classes['__root']).toBeDefined()
     expect(renderStyles).toHaveBeenCalledWith({ color: 'red' })
@@ -179,24 +155,15 @@ describe('resolveStylesAndClasses', () => {
 
   test('does not caches resolved styles for different props', () => {
     spyOn(componentStyles, 'root').and.callThrough()
-    const { resolvedStyles } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      () => '',
-      true,
-      'Test4',
-      theme,
-      { primary: true },
-    )
-    const { resolvedStyles: secondResolvedStyles } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      () => '',
-      true,
-      'Test4',
-      theme,
-      {},
-    )
+    const options = resolveStylesParam({
+      cacheEnabled: true,
+      displayName: 'Test4',
+      stylesProps: { primary: true },
+    })
+    const { resolvedStyles } = resolveStyles(options)
+
+    options.stylesProps = { primary: false }
+    const { resolvedStyles: secondResolvedStyles } = resolveStyles(options)
 
     expect(resolvedStyles.root).toMatchObject({ color: 'red' })
     expect(componentStyles.root).toHaveBeenCalledTimes(1)
@@ -206,24 +173,16 @@ describe('resolveStylesAndClasses', () => {
 
   test('does not cache classes for different props', () => {
     const renderStyles = jest.fn().mockReturnValue('a')
-    const { classes } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
+    const options = resolveStylesParam({
       renderStyles,
-      true,
-      'Test5',
-      theme,
-      { primary: true },
-    )
-    const { classes: secondClasses } = resolveStylesAndClasses(
-      componentStyles,
-      styleParam,
-      renderStyles,
-      true,
-      'Test5',
-      theme,
-      {},
-    )
+      cacheEnabled: true,
+      displayName: 'Test5',
+      stylesProps: { primary: true },
+    })
+    const { classes } = resolveStyles(options)
+
+    options.stylesProps = { primary: false }
+    const { classes: secondClasses } = resolveStyles(options)
 
     expect(classes['__root']).toBeDefined()
     expect(renderStyles).toHaveBeenCalledWith({ color: 'red' })
